Physiological data and transformation in Excel
After performing the clustering on the data receive from Kaggle, it became clear that the cluster did not have the right information to cluster based on height or weight, so I had to find the data myself. I copied the tables from the NBA, added them in to Power Query editor, corrected the height data type, appended the tables, transformed the data, and removed columns of data I had already received from Kaggle. The "All Players" worksheet will be exported as a CSV and merged with the Kaggle data.